Splitting CoreCanvasOptions and CoreTextureOptions again
Also removing the downscale property, as it makes most sense to
implement it fully in EngineImage and not in Core.
