+++ packages/fim/src/core/CoreCanvasOptions.ts
@@ -1,0 +1,7 @@
+// FIM - Fast Image Manipulation Library for Javascript
+// Copyright (c) Leo C. Singleton IV <leo@leosingleton.com>
+// See LICENSE in the project root for license information.
+
+/** Options for `CoreCanvas` and its derived classes */
+export interface CoreCanvasOptions {
+}
