+++ packages/fim/src/internal/index.ts
@@ -12,11 +12,12 @@ export { defaultEngineOptions } from '../api/FimEngineOptions';
 export { defaultImageOptions, mergeImageOptions } from '../api/FimImageOptions';
 export { CoreCanvas } from '../core/CoreCanvas';
 export { CoreCanvas2D } from '../core/CoreCanvas2D';
+export { CoreCanvasOptions } from '../core/CoreCanvasOptions';
 export { CoreCanvasWebGL, EventListenerType } from '../core/CoreCanvasWebGL';
 export { CoreMimeType } from '../core/CoreMimeType';
-export { CoreCanvasOptions, CoreOptions, CoreTextureOptions } from '../core/CoreOptions';
 export { CoreShader } from '../core/CoreShader';
 export { CoreTexture } from '../core/CoreTexture';
+export { CoreTextureOptions } from '../core/CoreTextureOptions';
 export { ImageSource } from '../core/types/ImageSource';
 export { RenderingContext2D } from '../core/types/RenderingContext2D';
 export { RenderingContextWebGL } from '../core/types/RenderingContextWebGL';
